Create selenium-test.py
Selenium is a test framework that you can test web applications in different browsers. It is free to use because it is an open source software. For test command files, we can use different software language like C#, Java, Python. In today, there are lots of companies like Google, Netflix, HubSpot which use Selenium for production.

Today, we are going to use Selenium as a WebDriver. First, install Selenium. During installation, make sure you also install the Firefox Webdriver. This also means that you need FireFox Browser installed on your computer.

Before we start, do not forget to get Selenium.